chore: move contract class structs to common
