+++ crates/blockifier/src/execution/contract_class_test.rs
@@ -3,9 +3,10 @@ use std::sync::Arc;
 
 use assert_matches::assert_matches;
 use cairo_lang_starknet_classes::NestedIntList;
+use common::contract_class::{ContractClassV1, ContractClassV1Inner};
 use rstest::rstest;
 
-use crate::execution::contract_class::{ContractClassV1, ContractClassV1Inner};
+use crate::execution::contract_class::ContractClassV1Ext;
 use crate::transaction::errors::TransactionExecutionError;
 
 #[rstest]
