+++ crates/blockifier/src/transaction/account_transactions_test.rs
@@ -3,6 +3,7 @@ use std::sync::Arc;
 
 use cairo_vm::types::builtin_name::BuiltinName;
 use cairo_vm::vm::runners::cairo_runner::ResourceTracker;
+use common::contract_class::{ContractClass, ContractClassV1};
 use pretty_assertions::assert_eq;
 use rstest::rstest;
 use starknet_api::core::{calculate_contract_address, ClassHash, ContractAddress, PatriciaKey};
@@ -26,7 +27,7 @@ use crate::abi::abi_utils::{
     selector_from_name,
 };
 use crate::context::BlockContext;
-use crate::execution::contract_class::{ContractClass, ContractClassV1};
+use crate::execution::contract_class::ContractClassV1Ext;
 use crate::execution::entry_point::EntryPointExecutionContext;
 use crate::execution::syscalls::SyscallSelector;
 use crate::fee::fee_utils::{get_fee_by_gas_vector, get_sequencer_balance_keys};
