+++ crates/common/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+edition.workspace = true
+license.workspace = true
+name = "common"
+repository.workspace = true
+version = "0.0.0"
+description = "Common types for the sequencer repository."
+
+[features]
+testing = []
+
+[dependencies]
+cairo-lang-casm.workspace = true
+cairo-lang-starknet-classes.workspace = true
+cairo-vm.workspace = true
+derive_more.workspace = true
+serde.workspace = true
+serde_json.workspace = true
+starknet-types-core.workspace = true
+starknet_api.workspace = true
