+++ crates/gateway/src/compilation_test.rs
@@ -1,8 +1,9 @@
 use assert_matches::assert_matches;
-use blockifier::execution::contract_class::ContractClass;
+use blockifier::execution::contract_class::ClassInfoExt;
 use cairo_lang_sierra_to_casm::compiler::CompilationError;
 use cairo_lang_starknet_classes::allowed_libfuncs::AllowedLibfuncsError;
 use cairo_lang_starknet_classes::casm_contract_class::StarknetSierraCompilationError;
+use common::contract_class::ContractClass;
 use mempool_test_utils::starknet_api_test_utils::declare_tx as rpc_declare_tx;
 use rstest::{fixture, rstest};
 use starknet_api::core::CompiledClassHash;
